chore(funnels): revert UDF bitfield widening
Keep this PR validation only: 32 steps stays the limit, so the UDF source, binaries, and dev schema stay untouched. Widening to UInt64 remains the path if more steps are ever needed.

Generated-By: PostHog Desktop
Task-Id: 5b91d4f1-f886-4bf7-bbb5-9c21db12a638

+++ funnel-udf/src/steps.rs
@@ -13,7 +13,7 @@ pub struct EnteredTimestamp {
     pub excluded: bool,
     pub timings: Vec<f64>,
     pub uuids: Vec<Uuid>,
-    pub steps: u64,
+    pub steps: u32,
 }
 
 #[derive(Debug, Clone, Deserialize)]
@@ -42,7 +42,7 @@ pub struct Result(
     pub PropVal,
     pub Vec<f64>,
     pub Vec<Vec<Uuid>>,
-    pub u64,
+    pub u32,
 );
 
 struct Vars {
